+++ R/deconvolute.R
@@ -32,6 +32,8 @@
 #' @param convert_bulk either "ref" to convert bulk RNA-Seq to scRNA-Seq scaling
 #'   using reference data or "qqmap" using quantile mapping of the bulk to
 #'   scRNA-Seq datasets, or "none" (or `FALSE`) for no conversion.
+#' @param lambda single numeric value of ridge parameter lambda. Only applied to
+#'   subclass deconvolution, not applied to cell group analysis. Experimental.
 #' @param check_comp logical, whether to analyse compensation values across
 #'   subclasses. See [plot_comp()].
 #' @param npass Number of passes. If `npass` set to 2 or more this activates
@@ -80,6 +82,10 @@
 #' normally distributed either, which probably explains the need for a very high
 #' cut-off. In practice the choice of settings seems to be dataset dependent.
 #' 
+#' The ridge parameter lambda, which adds L2 regularisation to the compensation
+#' (moment) matrix is provided only for experimental purposes. Any small benefit
+#' seems to be outweighed by varying other parameters especially `nsubclass`.
+#' 
 #' @returns A list object of S3 class 'deconv' containing:
 #'   \item{call}{the matched call}
 #'   \item{mk}{the original 'cellMarkers' class object}
@@ -95,14 +101,13 @@
 #'       \item `rawcomp`, the original unadjusted compensation matrix
 #'       \item `comp_amount`, the final values for the amount of compensation
 #'       across each cell subclass after adjustment to prevent negative values
+#'       \item `X`, the (weighted) model matrix
 #'       \item `residuals`, residuals, that is gene expression minus fitted 
 #'       values
 #'       \item `var.e`, variance of weighted residuals for each gene
 #'       \item `weights`, vector of weights
 #'       \item `resvar`, \eqn{s^2} the estimate of the gene expression variance 
 #'       for each sample
-#'       \item `se`, standard errors of cell counts
-#'       \item `hat`, diagonal elements of the hat matrix
 #'       \item `removed`, vector of outlying genes removed during successive 
 #'       passes
 #'   }}
@@ -116,7 +121,7 @@
 #'   within cell group percentages. The total percentage still adds to 100%.}
 #'   \item{comp_amount}{original argument `comp_amount`}
 #'   \item{comp_check}{optional list element returned when `check_comp = TRUE`}
-#' @seealso [cellMarkers()] [updateMarkers()] [rstudent.deconv()]
+#' @seealso [cellMarkers()] [updateMarkers()] [se()] [rstudent.deconv()]
 #'   [cooks.distance.deconv()]
 #' @author Myles Lewis
 #' @importFrom matrixStats colMins rowQuantiles rowVars
@@ -134,6 +139,7 @@ deconvolute <- function(mk, test,
                         use_filter = TRUE,
                         arith_mean = FALSE,
                         convert_bulk = FALSE,
+                        lambda = NULL,
                         check_comp = FALSE,
                         npass = 1,
                         outlier_method = c("var.e", "cooks", "rstudent"),
@@ -145,8 +151,10 @@ deconvolute <- function(mk, test,
   .call <- match.call()
   weight_method <- match.arg(weight_method, c("none", "equal"))
   outlier_method <- match.arg(outlier_method)
+  if (outlier_method != "var.e") se <- TRUE
   test <- as.matrix(test)
   if (any(test < 0)) stop("`test` contains negative values")
+  if (length(lambda) > 1) stop("lambda can only accept a single value")
   
   if (isTRUE(convert_bulk)) convert_bulk <- "ref"
   if (isFALSE(convert_bulk)) convert_bulk <- "none"
@@ -168,7 +176,7 @@ deconvolute <- function(mk, test,
     if (convert_bulk != "none") logtest <- bulk2scfun(logtest)
     gtest <- deconv_adjust(logtest, cellmat, group_comp_amount, weights = NULL,
                            adjust_comp, count_space, weight_method,
-                           verbose = verbose, resid = FALSE)
+                           lambda = NULL, verbose = verbose, resid = FALSE)
   } else {
     gtest <- NULL
   }
@@ -191,7 +199,7 @@ deconvolute <- function(mk, test,
   atest <- deconv_multipass(logtest2, cellmat, comp_amount, weights,
                             weight_method, adjust_comp, count_space, npass,
                             outlier_method, outlier_cutoff, outlier_quantile,
-                            verbose, cores)
+                            lambda, verbose, cores)
   
   # subclass nested within group output/percent
   if (!is.null(gtest)) {
@@ -199,22 +207,22 @@ deconvolute <- function(mk, test,
     output2 <- lapply(levels(mk$cell_table), function(i) {
       output1 <- gtest$output[, i]
       ind <- mk$cell_table == i
-      subclass_i <- atest$output[, ind, drop = FALSE]
+      subclass_i <- fixzero(atest$output[, ind, drop = FALSE])
       rs <- rowSums(subclass_i)
       subclass_i / rs * output1
     })
     nest_output <- do.call(cbind, output2)
-    nest_output[is.na(nest_output)] <- 0
+    nest_output[!is.finite(nest_output)] <- 0  # fix division by 0
     # subclass percent as nested percent of groups
     pc2 <- lapply(levels(mk$cell_table), function(i) {
       pc1 <- gtest$percent[, i]
       ind <- mk$cell_table == i
-      subclass_i <- atest$output[, ind, drop = FALSE]
+      subclass_i <- fixzero(atest$output[, ind, drop = FALSE])
       rs <- rowSums(subclass_i)
       subclass_i / rs * pc1
     })
     nest_percent <- do.call(cbind, pc2)
-    nest_percent[is.na(nest_percent)] <- 0
+    nest_percent[!is.finite(nest_percent)] <- 0  # fix division by 0
   } else nest_output <- nest_percent <- NULL
   
   out <- list(call = .call, mk = mk, subclass = atest, group = gtest,
@@ -223,8 +231,8 @@ deconvolute <- function(mk, test,
   if (convert_bulk == "qqmap") out$qqmap <- qqmap
   if (check_comp) {
     if (verbose) message("analysing compensation")
-    out$comp_check <- comp_check(logtest2, cellmat, comp_amount,
-                                 weights, weight_method, count_space, cores)
+    out$comp_check <- comp_check(logtest2, cellmat, comp_amount, weights,
+                                 weight_method, count_space, lambda, cores)
   }
   class(out) <- "deconv"
   out
@@ -234,9 +242,9 @@ deconvolute <- function(mk, test,
 deconv_multipass <- function(test, cellmat, comp_amount, weights, weight_method,
                              adjust_comp, count_space, npass,
                              outlier_method, outlier_cutoff, outlier_quantile,
-                             verbose, cores) {
+                             lambda, verbose, cores) {
   fit <- deconv_adjust(test, cellmat, comp_amount, weights,
-                       adjust_comp, count_space, weight_method,
+                       adjust_comp, count_space, weight_method, lambda,
                        cores, verbose)
   metric <- outlier_metric(fit, outlier_method, outlier_quantile, count_space)
   outlier <- metric > outlier_cutoff
@@ -252,23 +260,25 @@ deconv_multipass <- function(test, cellmat, comp_amount, weights, weight_method,
     i <- i +1
     remove_genes <- sort(metric[outlier], decreasing = TRUE)
     removed <- c(removed, remove_genes)
-    rem_names <- names(remove_genes)
-    if (length(remove_genes) > 20) {
-      rem_names <- c(rem_names[1:20],
-                    paste0("... [", length(remove_genes), " genes]"))
+    if (verbose) {
+      rem_names <- names(remove_genes)
+      if (length(remove_genes) > 20) {
+        rem_names <- c(rem_names[1:20],
+                       paste0("... [", length(remove_genes), " genes]"))
+      }
+      message("Pass ", i, " - removed ", paste(rem_names, collapse = ", "))
     }
-    if (verbose) message("Pass ", i, " - removed ",
-                         paste(rem_names, collapse = ", "))
     test <- test[!outlier, , drop = FALSE]
     cellmat <- cellmat[!outlier, , drop = FALSE]
     weights <- weights[!outlier]
     if (nrow(cellmat) < ncol(cellmat)) stop("insufficient genes")
     fit <- deconv_adjust(test, cellmat, comp_amount, weights,
-                         adjust_comp, count_space, weight_method,
+                         adjust_comp, count_space, weight_method, lambda,
                          cores, verbose)
     metric <- outlier_metric(fit, outlier_method, outlier_quantile, count_space)
     outlier <- metric > outlier_cutoff
   }
+  
   fit$removed <- removed
   fit
 }
@@ -290,8 +300,8 @@ outlier_metric <- function(fit, outlier_method, outlier_quantile, count_space) {
 
 deconv_adjust <- function(test, cellmat, comp_amount, weights,
                           adjust_comp, count_space,
-                          weight_method = "", cores = 1L, verbose = TRUE,
-                          resid = TRUE) {
+                          weight_method = "", lambda,
+                          cores = 1L, verbose = TRUE, resid = TRUE) {
   comp_amount <- rep_len(comp_amount, ncol(cellmat))
   names(comp_amount) <- colnames(cellmat)
   if (!identical(rownames(test), rownames(cellmat)))
@@ -316,7 +326,10 @@ deconv_adjust <- function(test, cellmat, comp_amount, weights,
   }
   
   m_itself <- dotprod(cellmat, cellmat)
-  atest <- deconv(test, cellmat, comp_amount, m_itself)
+  if (!is.null(lambda)) m_itself <- m_itself + diag(nrow(m_itself)) * lambda
+  rawcomp <- solve(m_itself)
+  test.cellmat <- dotprod(test, cellmat)
+  atest <- deconv(test.cellmat, comp_amount, m_itself, rawcomp)
   if (any(atest$output < 0)) {
     if (adjust_comp) {
       minout <- colMins(atest$output)
@@ -327,7 +340,7 @@ deconv_adjust <- function(test, cellmat, comp_amount, weights,
         f <- function(x) {
           newcomp <- comp_amount
           newcomp[wi] <- x
-          ntest <- quick_deconv(test, cellmat, newcomp, m_itself, wi)
+          ntest <- quick_deconv(test.cellmat, newcomp, m_itself, rawcomp, wi)
           min(ntest, na.rm = TRUE)^2
         }
         if (comp_amount[wi] == 0) return(0)
@@ -335,7 +348,7 @@ deconv_adjust <- function(test, cellmat, comp_amount, weights,
         xmin$minimum
       }, progress = verbose, mc.cores = cores)
       comp_amount[w] <- unlist(newcomps)
-      atest <- deconv(test, cellmat, comp_amount, m_itself)
+      atest <- deconv(test.cellmat, comp_amount, m_itself, rawcomp)
       # fix floating point errors
       if (any(z <- atest$output < 0)) {
         attr(atest$output, "min") <- min(atest$output)
@@ -345,7 +358,7 @@ deconv_adjust <- function(test, cellmat, comp_amount, weights,
     } else if (verbose) message("negative cell proportion projection detected")
   }
   if (resid) {
-    X <- cellmat
+    atest$X <- X <- cellmat
     atest$residuals <- r <- residuals_deconv(oldtest, oldcellmat, atest$output)
     # adjust residuals & X by gene weights
     if (!is.null(weights)) r <- r * weights
@@ -355,24 +368,15 @@ deconv_adjust <- function(test, cellmat, comp_amount, weights,
     rss <- colSums(r^2)
     rdf <- nrow(r) - ncol(X)
     atest$resvar <- rss/rdf
-    # residuals row variance
-    Lv <- colSums(X^2)
-    iXTX <- atest$compensation / Lv
-    XTXse <- crossprod(X, var.e * X)
-    # var(beta) = (X' X)^-1 (X diag(e^2) X') (X' X)^-1
-    atest$se <- sqrt(diag(iXTX %*% XTXse %*% t(iXTX)))
-    # calculate hat matrix
-    # X (X' X)^-1 X'
-    atest$hat <- diag(X %*% iXTX %*% t(X))
   }
   atest
 }
 
 
-deconv <- function(test, cellmat, comp_amount, m_itself) {
-  rawcomp <- solve(m_itself)
-  mixcomp <- solve(m_itself, t(comp_amount * diag(nrow(m_itself)) + (1-comp_amount) * t(m_itself)))
-  output <- dotprod(test, cellmat) %*% mixcomp
+deconv <- function(test.cellmat, comp_amount, m_itself, rawcomp) {
+  endcomp <- comp_amount * diag(nrow(m_itself)) + (1-comp_amount) * t(m_itself)
+  mixcomp <- tcrossprod(rawcomp, endcomp)
+  output <- test.cellmat %*% mixcomp
   percent <- output / rowSums(output) * 100
   
   list(output = output, percent = percent, spillover = m_itself,
@@ -380,40 +384,17 @@ deconv <- function(test, cellmat, comp_amount, m_itself) {
        comp_amount = comp_amount)
 }
 
-
-quick_deconv <- function(test, cellmat, comp_amount, m_itself, wi) {
-  mixcomp <- solve(m_itself, t(comp_amount * diag(nrow(m_itself)) + (1-comp_amount) * t(m_itself)))
-  dotprod(test, cellmat) %*% mixcomp[, wi, drop = FALSE]
-}
-
-
-approxfun.matrix <- function(x, FUN) {
-  if (is.data.frame(x)) x <- as.matrix(x)
-  if (is.matrix(x)) {
-    out <- FUN(as.vector(x))
-    out <- matrix(out, nrow = nrow(x), dimnames = dimnames(x))
-    return(out)
-  }
-  FUN(x)
-}
-
-#' @importFrom stats approxfun
-sc2bulk <- function(x) {
-  sc2bulkfun <- approxfun(x = celseqfit$celseq, y = celseqfit$pred.bulk,
-                          yleft = 0, rule = 2)
-  approxfun.matrix(x, sc2bulkfun)
-}
-
-
-bulk2sc <- function(x) {
-  bulk2scfun <- approxfun(x = celseqfit$pred.bulk, y = celseqfit$celseq,
-                          yleft = 0, rule = 2)
-  approxfun.matrix(x, bulk2scfun)
+# single column
+quick_deconv <- function(test.cellmat, comp_amount, m_itself, rawcomp, wi) {
+  endcomp <- (1-comp_amount[wi]) * m_itself[, wi]
+  endcomp[wi] <- endcomp[wi] + comp_amount[wi]
+  mixcomp <- rawcomp %*% endcomp
+  test.cellmat %*% mixcomp
 }
 
 
 comp_check <- function(test, cellmat, comp_amount, weights, weight_method,
-                       count_space, cores) {
+                       count_space, lambda, cores) {
   comp_amount <- rep_len(comp_amount, ncol(cellmat))
   names(comp_amount) <- colnames(cellmat)
   if (count_space) {
@@ -433,15 +414,25 @@ comp_check <- function(test, cellmat, comp_amount, weights, weight_method,
   px <- seq(0, 1, 0.05)
   
   m_itself <- dotprod(cellmat, cellmat)
+  if (!is.null(lambda)) m_itself <- m_itself + diag(nrow(m_itself)) * lambda
+  rawcomp <- solve(m_itself)
+  test.cellmat <- dotprod(test, cellmat)
   out <- mclapply(seq_len(ncol(cellmat)), function(i) {
     newcomp <- comp_amount
     vapply(px, function(ci) {
       newcomp[i] <- ci
-      ntest <- quick_deconv(test, cellmat, newcomp, m_itself, i)
+      ntest <- quick_deconv(test.cellmat, newcomp, m_itself, rawcomp, i)
       min(ntest, na.rm = TRUE)
     }, numeric(1))
   }, mc.cores = cores)
   names(out) <- colnames(cellmat)
   out$px <- px
   out
+}
+
+
+# avoid division by near 0
+fixzero <- function(x) {
+  x[abs(x) < sqrt(.Machine$double.eps)] <- 0
+  x
 }