+++ NEWS.md
@@ -1,6 +1,34 @@
 News
 =====
 
+# cellGeometry 0.6.1
+###### 13/01/2026
+* Fix division by 0 in `nest_output`
+
+## New features
+
+* Massive speed up of `deconvolute()`.
+* Add ridge parameter `lambda` to `deconvolute()` [experimental].
+* Add `resvar` metric to `tune_deconv()` to examine residual variance of bulk 
+gene expression.
+
+## Important change
+
+* `deconvolute()` no longer returns `se` and `hat` in the `subclass` list 
+element. Standard errors can be obtained with the new function `se()`.
+
+## Important bugfix
+
+* R 4.5.2 for macOS arm64 (Apple M* Macs) has switched to using a faster BLAS
+(vecLib?) by default, which causes errors with parallelisation in `mclapply()`.
+The problem is isolated to R 4.5.2 arm64 for macOS on M* Macs; or any version of
+R for macOS including intel, if vecLib BLAS is being used via a symlink. The
+solution is to use `cores = 1` with `deconvolute()` and `tune_deconv()` whenever
+vecLib BLAS is in use.
+* Removed use of `pbmcapply::pbmclapply()` as this caused problems with R 4.5.2 
+for macOS arm64 (Apple M* Macs) even with `cores=1`. This fixes indefinite 
+hanging in `tune_deconv()` associated with vecLib BLAS.
+
 # cellGeometry 0.5.7
 ###### 11/12/2025
 * Change `log` argument in `deconvolute()` to `logged_bulk`. NB. this is a